day07 add multithreading
Good little learning experience. In this case, seeing Polyester.jl be basically equivalent to Threads.@threads, so idk if that's bc base Julia has improved in the last couple years or what (most stuff related to Polyester seems to be from 2021/2). Either way, about 8x speedup on my desktop with auto threadcount of 24 for Julia, so that's neat (not sure why it's not even better tho...)