day12 p2
This one was a pain. Had to get a hint from reddit about looking for corners bc my approach for walking the edge of the shape was failing off-by-1 a lot but I couldn't think of a good way to fix it. In hindsight, maybe could have fixed it by saving row/col + direction instead of looking for repeated corners