day16 p2
Had a dumb issue with this bc I used a default dict at first for the prev path dict with lists, and the way i wrote it I guess every value was referencing the same list so all would get cleared nad overwritten (so at hte end, every node pointed to the previous node of the end point). Fixed that and it was fine.